feat: add mlflow s3 bucket name as config
Adds the "default_artifact_root" config value, which sets the s3 bucket that Argo Workflows uses as the default storage bucket (where it puts data if a Workflow does not specify its own bucket).  This is also useful in gateway mode, where you may use a bucket in an s3 store that already exists with a name.

fixes #24