fix: Pass ENV_AUTH value as string (#190)
According to the Seldon docs, `RCLONE_CONFIG_S3_ENV_AUTH` receives a
boolean value. However, passing `true` or `false` will cause the
metacontroller to break when attempting to convert the Secret from JSON
to a Go struct, since it's unable to unmarshall the boolean field. This
in turn, will lead to the secret not being applied.

Consequently, we have to quote the boolean value to ensure it's
interpreted as a string, e.g. `RCLONE_CONFIG_S3_ENV_AUTH: "false"`.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@canonical.com>